+++ python/sdss_access/path/path.py
@@ -6,6 +6,7 @@ import requests
 import ast
 import inspect
 import six
+import datetime
 from glob import glob
 from os.path import join, sep
 from random import choice, sample
@@ -18,10 +19,10 @@ try:
 except ImportError:
     import pathlib2 as pathlib
 
-try:
-    from ConfigParser import RawConfigParser
-except ImportError:
-    from configparser import RawConfigParser
+# try:
+#     from ConfigParser import RawConfigParser
+# except ImportError:
+#     from configparser import RawConfigParser
 
 """
 Module for constructing paths to SDSS files.
@@ -35,6 +36,40 @@ Example use case:
 Depends on the tree product. In particular requires path templates in:
   $TREE_DIR/data/sdss_paths.ini
 """
+
+
+def check_public_release(release: str = None, public: bool = False) -> bool:
+    """ Check if a release is public 
+    
+    Checks a given release to see if it is public.  A release is public if it
+    contains "DR" in the release name, and if todays date is <= the release_date
+    as specified in the Tree. 
+    
+    Parameters
+    ----------
+    release : str
+        The name of the release to check
+    public : bool
+        If True, force the release to be public
+
+    Returns
+    -------
+    bool
+        If the release if public
+        
+    Raises
+    ------
+    AttributeError
+        when tree does not have a valid release date for a DR tree config
+    """
+    today = datetime.datetime.now().date()
+    release_date = getattr(tree, 'release_date', None)
+    
+    # check if tree has a valid release date attr
+    if release_date is None and "DR" in tree.release:
+        raise AttributeError("Cannot find a valid release date in the sdss-tree product.  Try upgrading to min. version 3.1.0.")
+
+    return ('dr' in release.lower() and release_date <= today) or public
 
 
 class BasePath(object):
@@ -63,28 +98,34 @@ class BasePath(object):
         The set of templates read from the configuration file.
     """
 
-    _netloc = {"dtn": "sdss@dtn01.sdss.org", "sdss": "data.sdss.org",
-               "mirror": "data.mirror.sdss.org", 'svn': 'svn.sdss.org'}
+    _netloc = {"dtn": "dtn.sdss.org", "sdss": "data.sdss.org", "sdss5": "data.sdss5.org",
+               "mirror": "data.mirror.sdss.org", "svn": "svn.sdss.org"}
 
     def __init__(self, release=None, public=False, mirror=False, verbose=False,
                  force_modules=None, preserve_envvars=None):
+        # set release
         self.release = release or os.getenv('TREE_VER', 'sdsswork')
-        self.public = 'dr' in self.release.lower() or public
-        self.mirror = mirror
         self.verbose = verbose
         self.force_modules = force_modules or config.get('force_modules')
         self.preserve_envvars = preserve_envvars or config.get('preserve_envvars')
-        self.set_netloc()
-        self.set_remote_base()
 
+        # set attributes
         self._special_fxn_pattern = r"\@\w+[|]"
-
         self._compressions = ['.gz', '.bz2', '.zip']
         self._comp_regex = r'({0})$'.format('|'.join(self._compressions))
+
         # set the path templates from the tree
         self.templates = tree.paths
         if self.release:
             self.replant_tree(release=self.release)
+            
+        # set public and mirror keywords
+        self.public = check_public_release(release=self.release, public=public)
+        self.mirror = mirror
+            
+        # set the server location and remote base
+        self.set_netloc()
+        self.set_remote_base()
 
     def __repr__(self):
         return '<BasePath(release="{0}", public={1}, n_paths={2})'.format(self.release.lower(), self.public, len(self.templates))
@@ -740,7 +781,7 @@ class BasePath(object):
 
         return template
 
-    def get_netloc(self, netloc=None, sdss=None, dtn=None, svn=None, mirror=None):
+    def get_netloc(self, netloc=None, sdss=None, sdss5=None, dtn=None, svn=None, mirror=None):
         ''' Get a net url domain
 
         Returns an SDSS url domain location.  Options are the SDSS SAS domain, the rsync download
@@ -753,8 +794,10 @@ class BasePath(object):
                 An exact net location to return directly
             sdss : bool
                 If True, returns SDSS data domain: data.sdss.org
+            sdss5 : bool
+                If True, sets the SDSS-V data domain: data.sdss5.org
             dtn : bool
-                If True, returns SDSS rsync server domain: dtn01.sdss.org
+                If True, returns SDSS rsync server domain: dtn.sdss.org
             svn: bool
                 If True, returns SDSS svn domain: svn.sdss.org
             mirror: bool
@@ -770,15 +813,17 @@ class BasePath(object):
         if dtn:
             return self._netloc["dtn"]
         elif sdss:
-            return self._netloc['sdss']
+            return self._netloc["sdss"]
+        elif sdss5:
+            return self._netloc["sdss5"]
         elif mirror or self.mirror:
             return self._netloc["mirror"]
         elif svn:
-            return '{0}{1}'.format(self._netloc['svn'], '/public' if self.public else '')
+            return '{0}{1}'.format(self._netloc["svn"], "/public" if self.public else '')
         else:
-            return self._netloc['sdss']
+            return self._netloc["sdss5"] if self.release == "sdss5" else self._netloc["sdss"]
 
-    def set_netloc(self, netloc=None, sdss=None, dtn=None, svn=None, mirror=None):
+    def set_netloc(self, netloc=None, sdss=None, sdss5=None, dtn=None, svn=None, mirror=None):
         ''' Set a url domain location
 
         Sets an SDSS url domain location.  Options are the SDSS SAS domain, the rsync download
@@ -790,16 +835,18 @@ class BasePath(object):
             netloc : str
                 An exact net location to use directly
             sdss : bool
-                If True, sets the SDSS data domain: data.sdss.org
+                If True, sets the SDSS-IV data domain: data.sdss.org
+            sdss5 : bool
+                If True, sets the SDSS-V data domain: data.sdss5.org
             dtn : bool
-                If True, sets the SDSS rsync server domain: dtn01.sdss.org
+                If True, sets the SDSS rsync server domain: dtn.sdss.org
             svn: bool
                 If True, sets the SDSS svn domain: svn.sdss.org
             mirror: bool
                 If True, sets the SDSS mirror domain: data.mirror.sdss.org.
 
         '''
-        self.netloc = self.get_netloc(netloc=netloc, sdss=sdss, dtn=dtn, svn=svn, mirror=mirror)
+        self.netloc = self.get_netloc(netloc=netloc, sdss=sdss, sdss5=sdss5, dtn=dtn, svn=svn, mirror=mirror)
 
     def set_remote_base(self, scheme='https'):
         self.remote_base = self.get_remote_base(scheme=scheme or 'https')
@@ -817,7 +864,12 @@ class BasePath(object):
         netloc = self.netloc
         if svn:
             netloc = self.get_netloc(svn=True)
-        return "{scheme}://{netloc}".format(scheme=scheme, netloc=netloc)
+        if self.public or scheme == "https":
+            remote_base = "{scheme}://{netloc}".format(scheme=scheme, netloc=netloc)
+        else:
+            user = "sdss5" if self.release == "sdss5" else "sdss"
+            remote_base = "{scheme}://{user}@{netloc}".format(scheme=scheme, user=user, netloc=netloc)
+        return remote_base
 
     def set_base_dir(self, base_dir=None):
         ''' Sets the base directory