+++ python/sdss_access/sync/stream.py
@@ -18,7 +18,7 @@ class Stream(object):
             self.stream_count = min(int(stream_count), self.max_stream_count)
         except Exception:
             self.stream_count = 0
-        self.streamlet = [{'index': index, 'location': None, 'source': None, 'destination': None} for index in range(0, self.stream_count)]
+        self.streamlet = [{'index': index, 'sas_module': None, 'location': None, 'source': None, 'destination': None} for index in range(0, self.stream_count)]
         self.reset()
         self.index = 0
         self.command = None
@@ -36,19 +36,19 @@ class Stream(object):
 
     def reset_streamlet(self):
         for index in range(0, self.stream_count):
-            self.set_streamlet(index=index, location=[], source=[], destination=[])
+            self.set_streamlet(index=index, sas_module=[], location=[], source=[], destination=[])
 
-    def set_streamlet(self, index=None, location=None, source=None, destination=None):
+    def set_streamlet(self, index=None, sas_module=None, location=None, source=None, destination=None):
         streamlet = self.get_streamlet(index=index)
         if streamlet:
             try:
                 n = len(location)
-                ok = n == len(source) and n == len(destination)
-                streamlet['location'], streamlet['source'], streamlet['destination'] = (
-                    location, source, destination) if ok else (None, None, None)
+                ok = n == len(sas_module) and n == len(source) and n == len(destination)
+                streamlet['sas_module'], streamlet['location'], streamlet['source'], streamlet['destination'] = (
+                    sas_module, location, source, destination) if ok else (None, None, None, None)
             except Exception:
-                streamlet['location'], streamlet['source'], streamlet['destination'] = (
-                    None, None, None)
+                streamlet['sas_module'], streamlet['location'], streamlet['source'], streamlet['destination'] = (
+                    None, None, None, None)
 
     def get_streamlet(self, index=None, increment=1):
         if index is None:
@@ -87,9 +87,10 @@ class Stream(object):
         subset = filter(lambda i: r.search(i), locations)
         self.task = [self.task[locations.index(s)] for s in subset]
 
-    def append_task(self, location=None, source=None, destination=None):
-        if location and source and destination:
-            task = {'location': location, 'source': source, 'destination': destination, 'exists': None}
+    def append_task(self, sas_module=None, location=None, source=None, destination=None):
+        if sas_module and location and source and destination:
+            task = {'sas_module': sas_module, 'location': location, 'source': source,
+                    'destination': destination, 'exists': None}
             self.task.append(task)
 
     def append_tasks_to_streamlets(self, offset=None, limit=None):
@@ -107,6 +108,7 @@ class Stream(object):
     def append_streamlet(self, index=None, task=None):
         streamlet = self.get_streamlet(index=index)
         if streamlet and task:
+            streamlet['sas_module'].append(task['sas_module'])
             streamlet['location'].append(task['location'])
             streamlet['source'].append(task['source'])
             streamlet['destination'].append(task['destination'])
@@ -123,21 +125,27 @@ class Stream(object):
         if streamlet:
             streamlet['path'] = self.cli.get_path(index=streamlet['index'])
             path_txt = "{0}.txt".format(streamlet['path'])
-            streamlet['command'] = self.command.format(path=path_txt, source=self.source, destination=self.destination)
+            streamlet['location']
+            streamlet['command'] = self.command.format(path=path_txt, sas_module=self.sas_module,
+                                                        source=self.source, destination=self.destination)
 
             if 'rsync -' in self.command:
-                self.cli.write_lines(path=path_txt, lines=[location for location in streamlet['location']])
+                lines = [location for location in streamlet['location']]
             else:
                 if not is_posix:
-                    self.cli.write_lines(path=path_txt, lines=['url ' + join(self.source, location).replace(sep,'/')+'\n'+'output '+join(self.destination, location) for location in streamlet['location']])
+                    lines = ['url ' + join(self.source, location).replace(sep,'/')+'\n'+'output ' +
+                            join(self.destination, location) for location in streamlet['location']]
                 else:
-                    self.cli.write_lines(path=path_txt, lines=['url ' + join(self.source, location)+'\n'+'output '+join(self.destination, location) for location in streamlet['location']])
+                    lines = ['url ' + join(self.source, location)+'\n'+'output ' +
+                            join(self.destination, location) for location in streamlet['location']]
+            self.cli.write_lines(path=path_txt, lines=lines)
 
     def run_streamlets(self):
         for streamlet in self.streamlet:
             streamlet['logfile'] = open("{0}.log".format(streamlet['path']), "w")
             streamlet['errfile'] = open("{0}.err".format(streamlet['path']), "w")
-            streamlet['process'] = self.cli.get_background_process(streamlet['command'], logfile=streamlet['logfile'], errfile=streamlet['errfile'])
+            streamlet['process'] = self.cli.get_background_process(streamlet['command'],
+                                        logfile=streamlet['logfile'], errfile=streamlet['errfile'])
             if self.verbose:
                 print("SDSS_ACCESS> rsync stream %s logging to %s" % (streamlet['index'],streamlet['logfile'].name))